chore: remove codecov action from CI
CodeCov has been complaining about requiring a token for a *long* time. CodeCov isn't available on repos in the `edx` GitHub org, it is only available for repos/projects of the Open edX org (`openedx`).

This PR removes the step to try and generate a coverage report, which has just been failing anyway.

This also means it's one less dependency for Renovate to try and update.